+++ app/db/sqlalchemy/models/meal.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column
-from sqlalchemy.dialects.postgresql import VARCHAR, INTEGER
+from sqlalchemy.dialects.postgresql import VARCHAR, BIGINT
 from sqlalchemy import Table
 from app.db.sqlalchemy.base import metadata_obj, mapper_registry
 from app.core.models.meal import Meal
@@ -8,7 +8,7 @@ from app.core.models.meal import Meal
 meal = Table(
     "meal",
     metadata_obj,
-    Column("id", INTEGER, primary_key=True),
+    Column("id", BIGINT, primary_key=True),
     Column("name", VARCHAR)
 )
