+++ app/db/sqlalchemy/models/metric_type.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column
-from sqlalchemy.dialects.postgresql import VARCHAR, INTEGER
+from sqlalchemy.dialects.postgresql import VARCHAR, BIGINT
 from sqlalchemy import Table
 from app.db.sqlalchemy.base import metadata_obj, mapper_registry
 from app.core.models.metric_type import MetricType
@@ -8,7 +8,7 @@ from app.core.models.metric_type import MetricType
 metric_type = Table(
     "metric_type",
     metadata_obj,
-    Column("id", INTEGER, primary_key=True),
+    Column("id", BIGINT, primary_key=True),
     Column("type", VARCHAR),
     Column("description", VARCHAR)
 )
